requirements.txt
fastmcp==0.4.0
langgraph==0.2.46
langchain-core==0.3.34
langchain-community==0.3.14
fastapi==0.115.6
uvicorn==0.32.0
pydantic==2.10.3
python-dotenv==1.0.1
pytest==7.4.4
httpx==0.28.1
